move transaction model to common model layer
