+++ lib/features/transactions/domain/use_cases/create_transaction_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:shmr_finance/core/base/use_case_abstract/base_use_case.dart';
 
 import '../../data/dto/create_transaction_use_case_request.dart';
-import '../entities/transaction.dart';
+import '../../../../model/transaction.dart';
 import '../repositories/transactions_repository.dart';
 
 class CreateTransactionUseCase
